backend_flask/requirements.txt
Werkzeug <= 2.1.2   
flask == 2.1.2   
flask-restx >= 0.5.1
JPype1 == 1.4.0
konlpy == 0.6.0 
flask-cors == 3.0.10
wheel == 0.37.1
waitress == 2.0.0